+++ cli/cmd/get_event_approvalTriggered_test.go
@@ -59,8 +59,6 @@ const eventsResponse = `{
 }`
 
 func Test_getApprovalTriggeredEvents(t *testing.T) {
-
-	//mocking = true
 	ts := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			w.Header().Add("Content-Type", "application/json")
