+++ cli/cmd/root_test.go
@@ -148,7 +148,6 @@ func (r *redirector) revertStdErr() string {
 }
 
 func Test_runVersionCheck(t *testing.T) {
-	//mocking = true
 	var returnedMetadataStatus int
 	var returnedMetadata keptnapimodels.Metadata
