+++ cli/cmd/send_event_approvalFinished_test.go
@@ -119,8 +119,6 @@ const serviceResponse = `
         }`
 
 func Test_sendApprovalFinishedEvent(t *testing.T) {
-
-	//mocking = true
 	ts := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			w.Header().Add("Content-Type", "application/json")
