+++ cli/cmd/send_event_newArtifact_test.go
@@ -69,7 +69,6 @@ func TestNewArtifact(t *testing.T) {
 	checkEndPointStatusMock = true
 
 	receivedEvent := make(chan bool)
-	//mocking = true
 	ts := httptest.NewServer(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			w.Header().Add("Content-Type", "application/json")
