+++ cli/cmd/version_test.go
@@ -26,8 +26,6 @@ import (
 func TestVersionCmd(t *testing.T) {
 	credentialmanager.MockAuthCreds = true
 	checkEndPointStatusMock = true
-	//mocking = true
-
 	cmd := fmt.Sprintf("version")
 	Version = "0.6.1"
