+++ cli/internal/auth/store.go
@@ -8,7 +8,6 @@ import (
 	"golang.org/x/oauth2"
 	"io/ioutil"
 	"os"
-	"sync"
 )
 
 // Oauthinfo is a wrapper for oauth related information that is
@@ -223,109 +222,4 @@ func getDefaultClientValuesLocation() string {
 		return ClientValuesFileName
 	}
 	return configPath + ClientValuesFileName
-}
-
-// TokenStoreMock is an implementation of OauthStore usable as a mock in tests
-type TokenStoreMock struct {
-	sync.Mutex
-	storedOauthInfo       *OauthInfo
-	storedToken           *oauth2.Token
-	storedTokenDiscovery  *OauthDiscoveryResult
-	storedClientValues    *OauthClientValues
-	storeOauthInfoFn      func(*OauthInfo) error
-	getOauthInfoFn        func() (*OauthInfo, error)
-	getTokenFn            func() (*oauth2.Token, error)
-	storeTokenFn          func(*oauth2.Token) error
-	deleteTokenFn         func() error
-	getTokenDiscoveryFn   func() (*OauthDiscoveryResult, error)
-	storeTokenDiscoveryFn func(*OauthDiscoveryResult) error
-	storeClientValuesFn   func(values *OauthClientValues) error
-	getClientValuesFn     func() (*OauthClientValues, error)
-}
-
-func (t *TokenStoreMock) StoreOauthInfo(i *OauthInfo) error {
-	t.Lock()
-	defer t.Unlock()
-	if t.storeOauthInfoFn != nil {
-		return t.storeOauthInfoFn(i)
-	}
-	t.storedOauthInfo = i
-	t.storedToken = i.Token
-	t.storedTokenDiscovery = i.DiscoveryInfo
-	t.storedClientValues = i.ClientValues
-
-	return nil
-}
-
-func (t *TokenStoreMock) GetOauthInfo() (*OauthInfo, error) {
-	t.Lock()
-	defer t.Unlock()
-	if t.getOauthInfoFn != nil {
-		return t.getOauthInfoFn()
-	}
-	return t.storedOauthInfo, nil
-}
-
-func (t *TokenStoreMock) GetTokenInfo() (*oauth2.Token, error) {
-	t.Lock()
-	defer t.Unlock()
-	if t.getTokenFn != nil {
-		return t.getTokenFn()
-	}
-	return t.storedToken, nil
-}
-
-func (t *TokenStoreMock) StoreTokenInfo(token *oauth2.Token) error {
-	t.Lock()
-	defer t.Unlock()
-	if t.storeTokenFn != nil {
-		return t.storeTokenFn(token)
-	}
-	t.storedToken = token
-	return nil
-}
-func (t *TokenStoreMock) Wipe() error {
-	t.Lock()
-	defer t.Unlock()
-	if t.Wipe() != nil {
-		return t.deleteTokenFn()
-	}
-	return nil
-}
-
-func (t *TokenStoreMock) GetDiscoveryInfo() (*OauthDiscoveryResult, error) {
-	t.Lock()
-	defer t.Unlock()
-	if t.getTokenDiscoveryFn != nil {
-		return t.getTokenDiscoveryFn()
-	}
-	return t.storedTokenDiscovery, nil
-}
-func (t *TokenStoreMock) StoreDiscoveryInfo(discoveryResult *OauthDiscoveryResult) error {
-	t.Lock()
-	defer t.Unlock()
-	if t.storeTokenDiscoveryFn != nil {
-		return t.storeTokenDiscoveryFn(discoveryResult)
-	}
-	t.storedTokenDiscovery = discoveryResult
-	return nil
-}
-
-func (t *TokenStoreMock) StoreClientInfo(values *OauthClientValues) error {
-	t.Lock()
-	defer t.Unlock()
-	if t.storeClientValuesFn != nil {
-		return t.storeClientValuesFn(values)
-	}
-	t.storedClientValues = values
-	return nil
-}
-
-func (t *TokenStoreMock) GetClientInfo() (*OauthClientValues, error) {
-	t.Lock()
-	defer t.Unlock()
-	if t.getClientValuesFn != nil {
-		return t.getClientValuesFn()
-	}
-	return t.storedClientValues, nil
 }
